reorganize class hierarchies so base classes are closer to subclasses
