+++ nimare/meta/cbma/base.py
@@ -1,0 +1,7 @@
+from ...base.estimators import Estimator
+
+
+class CBMAEstimator(Estimator):
+    """Base class for coordinate-based meta-analysis methods.
+    """
+    pass
